Deprecated the original Isolated class, it is now replaced with using StatefulIsolate and TailoredStatefulIsolate directly
