+++ packages/integral_isolates/test/backpressure_tailored_isolated_test.dart
@@ -15,7 +15,7 @@ void main() {
     Future<List<int>> runIsolate(
       BackpressureStrategy<int, int> strategy,
     ) async {
-      final isolated = Isolated.tailored<int, int>(
+      final isolated = TailoredStatefulIsolate<int, int>(
         backpressureStrategy: strategy,
         autoInit: false,
       );
@@ -44,7 +44,7 @@ void main() {
     }
 
     test('Default (no) strategy', () async {
-      final isolate = Isolated();
+      final isolate = StatefulIsolate();
       final responses = <int>[];
 
       await Future.wait([
@@ -73,7 +73,7 @@ void main() {
     });
 
     test('Combine backpressure strategy', () async {
-      final isolate = Isolated.tailored<int, int>(
+      final isolate = TailoredStatefulIsolate<int, int>(
         backpressureStrategy: CombineBackPressureStrategy(
           (oldData, newData) => oldData + newData,
         ),
